feat: move data loading to App
